[pkg/networkdevice] Move GetFreePort to pkg/networkdevice
