+++ pkg/snmp/traps/server_test.go
@@ -6,6 +6,7 @@
 package traps
 
 import (
+	ndmtestutils "github.com/DataDog/datadog-agent/pkg/networkdevice/testutils"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -13,13 +14,13 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/aggregator/mocksender"
 )
 
-var freePort = getFreePort()
-
 func TestStartFailure(t *testing.T) {
 	/*
 		Start two servers with the same config to trigger an "address already in use" error.
 	*/
 
+	freePort, err := ndmtestutils.GetFreePort()
+	require.NoError(t, err)
 	config := Config{Port: freePort, CommunityStrings: []string{"public"}}
 	Configure(t, config)
